src/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aioredis==1.3.1
amqp==5.1.1; python_version >= '3.6'
anyio==4.0.0; python_version >= '3.8'
asgiref==3.3.4; python_version >= '3.6'
async-timeout==4.0.3; python_full_version <= '3.11.2'
attrs==23.1.0; python_version >= '3.7'
autobahn==23.6.2; python_version >= '3.9'
automat==22.10.0
bandit==1.7.5; python_version >= '3.7'
billiard==4.1.0; python_version >= '3.7'
celery==5.3.4; python_version >= '3.8'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
cfgv==3.4.0; python_version >= '3.8'
channels-redis==3.4.0; python_version >= '3.7'
channels==3.0.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_version >= '3.7'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
constantly==23.10.4; python_version >= '3.8'
contourpy==1.2.0; python_version >= '3.9'
coverage==7.3.2; python_version >= '3.8'
cryptography==41.0.5; python_version >= '3.7'
cycler==0.12.1; python_version >= '3.8'
daphne==3.0.2; python_version >= '3.6'
defusedxml==0.8.0rc2; python_version >= '3.6'
deprecated==1.2.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
distlib==0.3.7
django-braces==1.15.0
django-ckeditor==6.3.0
django-cors-headers==3.13.0; python_version >= '3.7'
django-js-asset==2.1.0; python_version >= '3.8'
django-modeltranslation==0.18.11
django-mptt==0.15.0; python_version >= '3.9'
django-multiurl==1.5.0
django-oauth-toolkit==2.3.0
django-phonenumber-field-for-garpix-user==8.0.0
django-polymorphic-tree-for-garpix-page==2.1.1
django-polymorphic==3.1.0
django-rest-framework-social-oauth2==1.1.0
django-solo==2.1.0; python_version >= '3.7'
django-tabbed-admin==1.0.4
django-tag-parser==3.2
django-uuslug==2.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
django==3.2.23
djangorestframework==3.14.0; python_version >= '3.6'
docutils==0.20.1; python_version >= '3.7'
drf-spectacular==0.26.5; python_version >= '3.6'
environs==9.5.0; python_version >= '3.6'
eqator==2.7.0
exceptiongroup==1.1.3; python_version < '3.11'
fcm-django==0.3.11
filelock==3.13.1; python_version >= '3.8'
flake8-polyfill==1.0.2
flake8==6.1.0; python_full_version >= '3.8.1'
fonttools==4.44.0; python_version >= '3.8'
future==0.18.3; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
garpix-admin-lock==1.2.0
garpix-auth==2.3.0
garpix-menu==1.16.0
garpix-notify==5.16.2
garpix-package==2.0.1
garpix-page==2.48.2
garpix-user==3.9.1
garpix-utils==1.9.0
garpixcms==4.5.2
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.40; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
hiredis==2.2.3; python_version >= '3.7'
httpcore==1.0.1; python_version >= '3.8'
httpx==0.25.1; python_version >= '3.8'
hyperlink==21.0.0
identify==2.5.31; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
importlib-metadata==6.8.0; python_version >= '3.8'
incremental==22.10.0
inflection==0.5.1; python_version >= '3.5'
jaraco.classes==3.3.0; python_version >= '3.8'
jsonschema-specifications==2023.7.1; python_version >= '3.8'
jsonschema==4.19.2; python_version >= '3.8'
jwcrypto==1.5.0; python_version >= '3.6'
keyring==24.2.0; python_version >= '3.8'
kiwisolver==1.4.5; python_version >= '3.7'
kombu==5.3.2; python_version >= '3.8'
m2r==0.2.1
mando==0.7.1
markdown-it-py==3.0.0; python_version >= '3.8'
marshmallow==3.20.1; python_version >= '3.8'
matplotlib==3.8.1
mccabe==0.7.0; python_version >= '3.6'
mdurl==0.1.2; python_version >= '3.7'
mistune==0.8.4
model-bakery==1.17.0; python_version >= '3.8'
more-itertools==10.1.0; python_version >= '3.8'
msgpack==1.0.7; python_version >= '3.8'
nh3==0.2.14
nodeenv==1.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
numpy==1.26.1
oauthlib==3.2.2; python_version >= '3.6'
opencv-python==4.8.1.78
packaging==23.2; python_version >= '3.7'
pandas==2.1.2
pbr==5.11.1; python_version >= '2.6'
phonenumbers==8.13.24
pillow==10.1.0; python_version >= '3.8'
pkginfo==1.9.6; python_version >= '3.6'
platformdirs==3.11.0; python_version >= '3.7'
pre-commit==3.5.0; python_version >= '3.8'
prompt-toolkit==3.0.39; python_version >= '3.7'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycodestyle==2.11.1; python_version >= '3.8'
pycparser==2.21
pyfcm==1.4.9
pyflakes==3.1.0; python_version >= '3.8'
pygments==2.16.1; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
pyopenssl==23.3.0
pyparsing==3.1.1; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
python-slugify==8.0.1; python_version >= '3.7'
python-telegram-bot==20.6; python_version >= '3.8'
python3-openid==3.2.0
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
radon==6.0.1
rarfile==4.1; python_version >= '3.6'
readme-renderer==42.0; python_version >= '3.8'
redis==5.0.1; python_version >= '3.7'
referencing==0.30.2; python_version >= '3.8'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.31.0; python_version >= '3.7'
rfc3986==2.0.0; python_version >= '3.7'
rich==13.6.0; python_version >= '3.7'
rpds-py==0.12.0; python_version >= '3.8'
sentry-sdk==1.34.0
service-identity==23.1.0
setuptools==68.2.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
sniffio==1.3.0; python_version >= '3.7'
social-auth-app-django==5.1.0; python_version >= '3.7'
social-auth-core==4.3.0; python_version >= '3.6'
sqlparse==0.4.4; python_version >= '3.5'
stevedore==5.1.0; python_version >= '3.8'
text-unidecode==1.3
tqdm==4.66.1; python_version >= '3.7'
twilio==7.10.0; python_version >= '3.6'
twine==3.4.1; python_version >= '3.6'
twisted[tls]==23.10.0; python_version >= '3.8'
txaio==23.1.1; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
tzdata==2023.3; python_version >= '2'
unrar==0.4
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.0.7; python_version >= '3.6'
uwsgi==2.0.23
viberbot==1.0.12
vine==5.0.0; python_version >= '3.6'
virtualenv==20.24.6; python_version >= '3.7'
wcwidth==0.2.9
wrapt>=1.11.0,<2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zipp==3.17.0; python_version >= '3.8'
zope-interface==6.1; python_version >= '3.7'
scikit-learn==1.3.2
tensorflow==2.15.0
keras<2.16,>=2.15.0
openpyxl==3.1.2